requirements.txt
keras==2.9.0
Keras-Preprocessing
matplotlib
mintpy
numpy
pandas
scikit-learn==1.3.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorflow-metal==1.0.1
Theano==1.0.5
